Rename Rover Position class to Position Class
A position in itself can have x & y coordinates(in terms of being applicable to a two dimensional grid system) along with an orientation(i.e., one of the 4 cardinal compass points) and the ability for it to be applied to a given terrain, hence a generic Position class feels more appropriate here than a specific RoverPosition class.

Also it makes sense that a Rover is composed of a Position and it needn't only have to be restricted to a Rover is composed of a RoverPosition and also in retrospect, the former sounds better.